Move table to summary section
