+++ app/javascript/app/pages/mitigation/mitigation-effects/summary/summary-selectors.js
@@ -89,6 +89,39 @@ export const getVisTypeSelected = createSelector(
   }
 );
 
+const getEffectsIndicatorName = createSelector(
+  [ getSummaryData, getSummaryMeta ],
+  (data, meta) => {
+    if (!meta || isEmpty(meta) || !data || isEmpty(data)) return null;
+    const effectNames = {};
+    Object.keys(data[0]).forEach(key => {
+      if (key.startsWith('effects')) {
+        effectNames[key] = meta.find(m => m.code === key).indicator;
+      }
+    });
+    return effectNames;
+  }
+);
+
+const getTableData = createSelector(
+  [ getSummaryData, getSummaryMeta, getEffectsIndicatorName ],
+  (data, meta, effectNames) => {
+    if (!meta || isEmpty(meta) || !effectNames) return null;
+    const tableData = data.map(d => {
+      const updatedD = {};
+      Object.keys(d).forEach(key => {
+        if (key.startsWith('effects')) {
+          updatedD[effectNames[key]] = d[key];
+        } else {
+          updatedD[key] = d[key];
+        }
+      });
+      return updatedD;
+    });
+    return { data: tableData, defaultColumns: Object.keys(tableData[0]) };
+  }
+);
+
 export const getGHGSelected = createSelector([ getGHGOptions, getGHGParam ], (
   options,
   selected
@@ -164,6 +197,6 @@ export const getSummary = createStructuredSelector({
   visTypeSelected: getVisTypeSelected,
   GHGOptions: getGHGOptions,
   GHGSelected: getGHGSelected,
-  tableData: getSummaryData,
+  tableData: getTableData,
   query: getQueryParams
 });
